First attempt on indentation fixes

+++ upload/admin/model/extension/payment/eway.php
@@ -10,8 +10,8 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function install(): void {
-        $this->db->query("
+	public function install(): void {
+		$this->db->query("
 			CREATE TABLE IF NOT EXISTS `" . DB_PREFIX . "eway_order` (
 			  `eway_order_id` int(11) NOT NULL AUTO_INCREMENT,
 			  `order_id` int(11) NOT NULL,
@@ -27,7 +27,7 @@ class ModelExtensionPaymentEway extends Model {
 			  PRIMARY KEY (`eway_order_id`)
 			) ENGINE=MyISAM DEFAULT COLLATE=utf8_general_ci;");
 
-        $this->db->query("
+		$this->db->query("
 			CREATE TABLE IF NOT EXISTS `" . DB_PREFIX . "eway_transactions` (
 			  `eway_order_transaction_id` int(11) NOT NULL AUTO_INCREMENT,
 			  `eway_order_id` int(11) NOT NULL,
@@ -38,7 +38,7 @@ class ModelExtensionPaymentEway extends Model {
 			  PRIMARY KEY (`eway_order_transaction_id`)
 			) ENGINE=MyISAM DEFAULT COLLATE=utf8_general_ci;");
 
-        $this->db->query("
+		$this->db->query("
 			CREATE TABLE IF NOT EXISTS `" . DB_PREFIX . "eway_card` (
 			  `card_id` INT(11) NOT NULL AUTO_INCREMENT,
 			  `customer_id` INT(11) NOT NULL,
@@ -49,20 +49,20 @@ class ModelExtensionPaymentEway extends Model {
 			  `type` VARCHAR(50) NOT NULL,
 			  PRIMARY KEY (`card_id`)
 			) ENGINE=MyISAM DEFAULT COLLATE=utf8_general_ci;");
-    }
+	}
 
 	/**
 	 * Uninstall
 	 *
 	 * @return void
 	 */
-    public function uninstall(): void {
-        //$this->model_setting_setting->deleteSetting($this->request->get['extension']);
+	public function uninstall(): void {
+		//$this->model_setting_setting->deleteSetting($this->request->get['extension']);
 
-        $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "eway_order`;");
-        $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "eway_transactions`;");
-        $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "eway_card`;");
-    }
+		$this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "eway_order`;");
+		$this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "eway_transactions`;");
+		$this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "eway_card`;");
+	}
 
 	/**
 	 * getOrder
@@ -71,19 +71,19 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return array
 	 */
-    public function getOrder(int $order_id): array {
-        $query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "eway_order` WHERE `order_id` = '" . (int)$order_id . "' LIMIT 1");
+	public function getOrder(int $order_id): array {
+		$query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "eway_order` WHERE `order_id` = '" . (int)$order_id . "' LIMIT 1");
 
-        if ($query->num_rows) {
-            $order = $query->row;
+		if ($query->num_rows) {
+			$order = $query->row;
 
-            $order['transactions'] = $this->getTransactions($order['eway_order_id']);
+			$order['transactions'] = $this->getTransactions($order['eway_order_id']);
 
-            return $order;
-        } else {
-            return [];
-        }
-    }
+			return $order;
+		} else {
+			return [];
+		}
+	}
 
 	/**
 	 * addRefundRecord
@@ -93,19 +93,19 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function addRefundRecord(array $order, object $result): void {
-        $transaction_id = $result->TransactionID;
-        $total_amount = $result->Refund->TotalAmount / 100;
-        $refund_amount = $order['refund_amount'] + $total_amount;
+	public function addRefundRecord(array $order, object $result): void {
+		$transaction_id = $result->TransactionID;
+		$total_amount = $result->Refund->TotalAmount / 100;
+		$refund_amount = $order['refund_amount'] + $total_amount;
 
-        if (isset($order['refund_transaction_id']) && $order['refund_transaction_id'] != '') {
-            $order['refund_transaction_id'] .= ',';
-        }
+		if (isset($order['refund_transaction_id']) && $order['refund_transaction_id'] != '') {
+			$order['refund_transaction_id'] .= ',';
+		}
 
-        $order['refund_transaction_id'] .= $transaction_id;
+		$order['refund_transaction_id'] .= $transaction_id;
 
-        $this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `modified` = NOW(), `refund_amount` = '" . (double)$refund_amount . "', `refund_transaction_id` = '" . $this->db->escape($order['refund_transaction_id']) . "' WHERE `eway_order_id` = '" . $order['eway_order_id'] . "'");
-    }
+		$this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `modified` = NOW(), `refund_amount` = '" . (float)$refund_amount . "', `refund_transaction_id` = '" . $this->db->escape($order['refund_transaction_id']) . "' WHERE `eway_order_id` = '" . $order['eway_order_id'] . "'");
+	}
 
 	/**
 	 * Capture
@@ -116,29 +116,29 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return object|null
 	 */
-    public function capture(int $order_id, float $capture_amount, array $currency): ?object {
-        $eway_order = $this->getOrder($order_id);
+	public function capture(int $order_id, float $capture_amount, array $currency): ?object {
+		$eway_order = $this->getOrder($order_id);
 
-        if ($eway_order && $capture_amount > 0) {
-            $capture_data = new \stdClass();
-            $capture_data->Payment = new \stdClass();
-            $capture_data->Payment->TotalAmount = (int)(number_format($capture_amount, 2, '.', '') * 100);
-            $capture_data->Payment->CurrencyCode = $currency;
-            $capture_data->TransactionID = $eway_order['transaction_id'];
+		if ($eway_order && $capture_amount > 0) {
+			$capture_data = new \stdClass();
+			$capture_data->Payment = new \stdClass();
+			$capture_data->Payment->TotalAmount = (int)(number_format($capture_amount, 2, '.', '') * 100);
+			$capture_data->Payment->CurrencyCode = $currency;
+			$capture_data->TransactionID = $eway_order['transaction_id'];
 
-            if ($this->config->get('payment_eway_test')) {
-                $url = 'https://api.sandbox.ewaypayments.com/CapturePayment';
-            } else {
-                $url = 'https://api.ewaypayments.com/CapturePayment';
-            }
+			if ($this->config->get('payment_eway_test')) {
+				$url = 'https://api.sandbox.ewaypayments.com/CapturePayment';
+			} else {
+				$url = 'https://api.ewaypayments.com/CapturePayment';
+			}
 
-            $response = $this->sendCurl($url, $capture_data);
+			$response = $this->sendCurl($url, $capture_data);
 
-            return json_decode($response);
-        } else {
-            return null;
-        }
-    }
+			return json_decode($response);
+		} else {
+			return null;
+		}
+	}
 
 	/**
 	 * updateCaptureStatus
@@ -148,9 +148,9 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function updateCaptureStatus(int $eway_order_id, int $status): void {
-        $this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `capture_status` = '" . (int)$status . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
-    }
+	public function updateCaptureStatus(int $eway_order_id, int $status): void {
+		$this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `capture_status` = '" . (int)$status . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
+	}
 
 	/**
 	 * updateTransactionId
@@ -160,9 +160,9 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function updateTransactionId(int $eway_order_id, string $transaction_id): void {
-        $this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `transaction_id` = '" . $this->db->escape($transaction_id) . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
-    }
+	public function updateTransactionId(int $eway_order_id, string $transaction_id): void {
+		$this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `transaction_id` = '" . $this->db->escape($transaction_id) . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
+	}
 
 	/**
 	 * Void
@@ -171,26 +171,26 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return object
 	 */
-    public function void(int $order_id): ?object {
-        $eway_order = $this->getOrder($order_id);
+	public function void(int $order_id): ?object {
+		$eway_order = $this->getOrder($order_id);
 
-        if ($eway_order) {
-            $data = new \stdClass();
-            $data->TransactionID = $eway_order['transaction_id'];
+		if ($eway_order) {
+			$data = new \stdClass();
+			$data->TransactionID = $eway_order['transaction_id'];
 
-            if ($this->config->get('payment_eway_test')) {
-                $url = 'https://api.sandbox.ewaypayments.com/CancelAuthorisation';
-            } else {
-                $url = 'https://api.ewaypayments.com/CancelAuthorisation';
-            }
+			if ($this->config->get('payment_eway_test')) {
+				$url = 'https://api.sandbox.ewaypayments.com/CancelAuthorisation';
+			} else {
+				$url = 'https://api.ewaypayments.com/CancelAuthorisation';
+			}
 
-            $response = $this->sendCurl($url, $data);
+			$response = $this->sendCurl($url, $data);
 
-            return json_decode($response);
-        } else {
-            return new \stdclass();
-        }
-    }
+			return json_decode($response);
+		} else {
+			return new \stdclass();
+		}
+	}
 
 	/**
 	 * updateVoidStatus
@@ -200,9 +200,9 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function updateVoidStatus(int $eway_order_id, int $status): void {
-        $this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `void_status` = '" . (int)$status . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
-    }
+	public function updateVoidStatus(int $eway_order_id, int $status): void {
+		$this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `void_status` = '" . (int)$status . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
+	}
 
 	/**
 	 * Refund
@@ -212,29 +212,29 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return string
 	 */
-    public function refund(int $order_id, float $refund_amount): string {
-        $eway_order = $this->getOrder($order_id);
+	public function refund(int $order_id, float $refund_amount): string {
+		$eway_order = $this->getOrder($order_id);
 
-        if ($eway_order && $refund_amount > 0) {
-            $refund_data = new \stdClass();
-            $refund_data->Refund = new \stdClass();
-            $refund_data->Refund->TotalAmount = (int)(number_format($refund_amount, 2, '.', '') * 100);
-            $refund_data->Refund->TransactionID = $eway_order['transaction_id'];
-            $refund_data->Refund->CurrencyCode = $eway_order['currency_code'];
+		if ($eway_order && $refund_amount > 0) {
+			$refund_data = new \stdClass();
+			$refund_data->Refund = new \stdClass();
+			$refund_data->Refund->TotalAmount = (int)(number_format($refund_amount, 2, '.', '') * 100);
+			$refund_data->Refund->TransactionID = $eway_order['transaction_id'];
+			$refund_data->Refund->CurrencyCode = $eway_order['currency_code'];
 
-            if ($this->config->get('payment_eway_test')) {
-                $url = 'https://api.sandbox.ewaypayments.com/Transaction/' . $eway_order['transaction_id'] . '/Refund';
-            } else {
-                $url = 'https://api.ewaypayments.com/Transaction/' . $eway_order['transaction_id'] . '/Refund';
-            }
+			if ($this->config->get('payment_eway_test')) {
+				$url = 'https://api.sandbox.ewaypayments.com/Transaction/' . $eway_order['transaction_id'] . '/Refund';
+			} else {
+				$url = 'https://api.ewaypayments.com/Transaction/' . $eway_order['transaction_id'] . '/Refund';
+			}
 
-            $response = $this->sendCurl($url, $refund_data);
+			$response = $this->sendCurl($url, $refund_data);
 
-            return json_decode($response);
-        } else {
-            return '';
-        }
-    }
+			return json_decode($response);
+		} else {
+			return '';
+		}
+	}
 
 	/**
 	 * updateRefundStatus
@@ -244,9 +244,9 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function updateRefundStatus(int $eway_order_id, int $status): void {
-        $this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `refund_status` = '" . (int)$status . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
-    }
+	public function updateRefundStatus(int $eway_order_id, int $status): void {
+		$this->db->query("UPDATE `" . DB_PREFIX . "eway_order` SET `refund_status` = '" . (int)$status . "' WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
+	}
 
 	/**
 	 * sendCurl
@@ -256,51 +256,51 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function sendCurl(string $url, array $data): string {
-        $ch = curl_init($url);
+	public function sendCurl(string $url, array $data): string {
+		$ch = curl_init($url);
 
-        $eway_username = html_entity_decode($this->config->get('payment_eway_username'), ENT_QUOTES, 'UTF-8');
-        $eway_password = html_entity_decode($this->config->get('payment_eway_password'), ENT_QUOTES, 'UTF-8');
+		$eway_username = html_entity_decode($this->config->get('payment_eway_username'), ENT_QUOTES, 'UTF-8');
+		$eway_password = html_entity_decode($this->config->get('payment_eway_password'), ENT_QUOTES, 'UTF-8');
 
-        curl_setopt($ch, CURLOPT_HTTPHEADER, ['Content-Type: application/json']);
-        curl_setopt($ch, CURLOPT_USERPWD, $eway_username . ':' . $eway_password);
-        curl_setopt($ch, CURLOPT_POST, 1);
-        curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($data));
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-        curl_setopt($ch, CURLOPT_TIMEOUT, 60);
-        curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 1);
+		curl_setopt($ch, CURLOPT_HTTPHEADER, ['Content-Type: application/json']);
+		curl_setopt($ch, CURLOPT_USERPWD, $eway_username . ':' . $eway_password);
+		curl_setopt($ch, CURLOPT_POST, 1);
+		curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($data));
+		curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+		curl_setopt($ch, CURLOPT_TIMEOUT, 60);
+		curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 1);
 
-        $response = curl_exec($ch);
+		$response = curl_exec($ch);
 
-        if (curl_errno($ch) != CURLE_OK) {
-            $response = new \stdClass();
+		if (curl_errno($ch) != CURLE_OK) {
+			$response = new \stdClass();
 
-            $response->Errors = "POST Error: " . curl_error($ch) . " URL: $url";
-            $response = json_encode($response);
-        } else {
-            $info = curl_getinfo($ch);
+			$response->Errors = "POST Error: " . curl_error($ch) . " URL: $url";
+			$response = json_encode($response);
+		} else {
+			$info = curl_getinfo($ch);
 
-            if ($info['http_code'] == 401 || $info['http_code'] == 404) {
-                $response = new \stdClass();
-                $response->Errors = "Please check the API Key and Password";
-                $response = json_encode($response);
-            }
-        }
+			if ($info['http_code'] == 401 || $info['http_code'] == 404) {
+				$response = new \stdClass();
+				$response->Errors = "Please check the API Key and Password";
+				$response = json_encode($response);
+			}
+		}
 
-        curl_close($ch);
+		curl_close($ch);
 
-        return $response;
-    }
+		return $response;
+	}
 
-    private function getTransactions(int $eway_order_id): array {
-        $query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "eway_transactions` WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
+	private function getTransactions(int $eway_order_id): array {
+		$query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "eway_transactions` WHERE `eway_order_id` = '" . (int)$eway_order_id . "'");
 
-        if ($query->num_rows) {
-            return $query->rows;
-        } else {
-            return [];
-        }
-    }
+		if ($query->num_rows) {
+			return $query->rows;
+		} else {
+			return [];
+		}
+	}
 
 	/**
 	 * addTransaction
@@ -313,9 +313,9 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return void
 	 */
-    public function addTransaction(int $eway_order_id, string $transactionid, string $type, float $total, string $currency): void {
-        $this->db->query("INSERT INTO `" . DB_PREFIX . "eway_transactions` SET `eway_order_id` = '" . (int)$eway_order_id . "', `created` = NOW(), `transaction_id` = '" . $this->db->escape($transactionid) . "', `type` = '" . $this->db->escape($type) . "', `amount` = '" . $this->currency->format($total, $currency, false, false) . "'");
-    }
+	public function addTransaction(int $eway_order_id, string $transactionid, string $type, float $total, string $currency): void {
+		$this->db->query("INSERT INTO `" . DB_PREFIX . "eway_transactions` SET `eway_order_id` = '" . (int)$eway_order_id . "', `created` = NOW(), `transaction_id` = '" . $this->db->escape($transactionid) . "', `type` = '" . $this->db->escape($type) . "', `amount` = '" . $this->currency->format($total, $currency, false, false) . "'");
+	}
 
 	/**
 	 * getTotalCaptured
@@ -324,11 +324,11 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return float
 	 */
-    public function getTotalCaptured(int $eway_order_id): float {
-        $query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "eway_transactions` WHERE `eway_order_id` = '" . (int)$eway_order_id . "' AND `type` = 'payment' ");
+	public function getTotalCaptured(int $eway_order_id): float {
+		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "eway_transactions` WHERE `eway_order_id` = '" . (int)$eway_order_id . "' AND `type` = 'payment' ");
 
-        return (float)$query->row['total'];
-    }
+		return (float)$query->row['total'];
+	}
 
 	/**
 	 * getTotalRefunded
@@ -337,9 +337,9 @@ class ModelExtensionPaymentEway extends Model {
 	 *
 	 * @return float
 	 */
-    public function getTotalRefunded(int $eway_order_id): float {
-        $query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "eway_transactions` WHERE `eway_order_id` = '" . (int)$eway_order_id . "' AND `type` = 'refund'");
+	public function getTotalRefunded(int $eway_order_id): float {
+		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "eway_transactions` WHERE `eway_order_id` = '" . (int)$eway_order_id . "' AND `type` = 'refund'");
 
-        return (float)$query->row['total'];
-    }
+		return (float)$query->row['total'];
+	}
 }
